% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLink.R
\name{getplinkinfo}
\alias{getplinkinfo}
\title{Title}
\usage{
getplinkinfo(famfile, bimfile, bedfile)
}
\arguments{
\item{famfile}{name of plink family file}

\item{bimfile}{name of plink binary map file}

\item{bedfile}{anme of plink binary SNP file}
}
\value{
List with information about binary plink data set
}
\description{
Title
}
\examples{
famfile <- system.file("extdata", "test.fam", package = "PlinkBin")
bimfile <- system.file("extdata", "test.bim", package = "PlinkBin")
bedfile <- system.file("extdata", "test.bed", package = "PlinkBin")
plinkinfo <- getplinkinfo(famfile, bimfile, bedfile)
}
