% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLink.R
\name{getplinksnp}
\alias{getplinksnp}
\title{Title}
\usage{
getplinksnp(plinkinfo, snp)
}
\arguments{
\item{plinkinfo}{Information about a binary plink data set returned from
getplinkinof}

\item{snp}{SNP name or number}
}
\value{
Integer vector with SNP values
}
\description{
Title
}
\examples{
famfile <- system.file("extdata", "test.fam", package = "LinGxEScanR")
bimfile <- system.file("extdata", "test.bim", package = "LinGxEScanR")
bedfile <- system.file("extdata", "test.bed", package = "LinGxEScanR")
plinkinfo <- getplinkinfo(famfile, bimfile, bedfile)

snpvalue <- getplinksnp(plinkinfo, "snp2")
}
