\name{LinGxEScanR-package}
\alias{LinGxEScanR-package}
\alias{LinGxEScanR}
\docType{package}
\title{
\packageTitle{LinGxEScanR}
}
\description{
\packageDescription{LinGxEScanR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{LinGxEScanR}
\packageIndices{LinGxEScanR}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{LinGxEScanR}

Maintainer: \packageMaintainer{LinGxEScanR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
