% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLink.R
\name{plinkapply2}
\alias{plinkapply2}
\title{Apply a function to each SNP in a
plink binary file}
\usage{
plinkapply2(plinkinfo, func, snps, ...)
}
\arguments{
\item{plinkinfo}{List with information about the
binary dosage file returned from getplinkinfo}

\item{func}{A user supplied function to apply
to the data for each snp. The function must be
provide with the following parameters, dosage,
p0, p1, and p2, where dosage is the dosage values
for each subject and p0, p1, and p2 are the
probabilities that a subject has zero, one,
and two copies of the alternate allele,
respectively.}

\item{snps}{List of SNPs to perform the function
on. Null value indicates to perform function all
all SNPs. Default NULL.}

\item{...}{Additional parameters needed by the
user supplied function}
}
\value{
A list with length equal to the number
of SNPs in the binary dosage file. Each element
of the list is the value returned by the user
supplied function
}
\description{
A routine that reads in the SNP data serially
from a binary dosage file and applies a user
specified function to the data.
}
\examples{
# Get information about a plink binary file

famfile <- system.file("extdata", "test.fam", package = "LinGxEScanR")
bimfile <- system.file("extdata", "test.bim", package = "LinGxEScanR")
bedfile <- system.file("extdata", "test.bed", package = "LinGxEScanR")
plinkinfo <- getplinkinfo(famfile, bimfile, bedfile)

# Apply the getaaf, get alternate allele frequency, function
# to all the SNPs in the binary dosage file

aaf <- plinkapply2(plinkinfo = plinkinfo,
               func = BinaryDosage:::getaaf)
}
\seealso{
Other Iterating functions: 
\code{\link{bdapply2}()},
\code{\link{vcfapply2}()}
}
\concept{Iterating functions}
