% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinGxEScanR.R
\name{lingweis}
\alias{lingweis}
\title{lingweis}
\usage{
lingweis(
  data,
  ginfo,
  snps,
  gonly,
  ge,
  ggxe,
  gxe,
  twodf,
  levene,
  outfile,
  skipfile,
  minmaf,
  blksize
)
}
\arguments{
\item{data}{Data frame containing the subject ID, phenotype
and covariates. The first column must be a character value that
contains the subject ID. The second column must be a numeric value and
contain the outcome. The remaining columns must be numeric and contain
the covariates to use in the analysis. The value in the last column is
the covariate that will be used in the gene-environment interaction.}

\item{ginfo}{Information about the binary dosage or vcf file returned
from the BinaryDosage::getbdinfo or BinaryDosage::getvcfinfo routine}

\item{snps}{The SNPs to be used in the scan. This may be an integer
vector indicate which SNPs to use in the binary dosage file or a
character vector of the SNP IDs to use. The value may also be "all",
indicating to use all SNPs. The default value is "all".}

\item{gonly}{The tests to perform on the beta_g parametetfor the model
with the gene and all the covariates except the one used in the
gene-environement interaction. The value must be a character array
containing any or all of the following values, "fit", "lrt", "score",
"Wald", and "WaldHW". This value can also be "all" or "none".}

\item{ge}{The tests to perform on the beta_g parameter from the model
with the all the covariates and the gene. The value has the sames
requirements as the gonly value.}

\item{ggxe}{The test to perform on the beta_g parameter from the model
that contains all the covariates, the gene, and the gene-environment
interaction. The values has the same requirements as the gonly value.}

\item{gxe}{The test to perform on the beta_gxe parameter from the model
that contains all the covariates, the gene, and the gene-environment
interaction. The values has the same requirements as the gonly value.}

\item{twodf}{The test two degree of tests to perform on the beta_g and
beta_gxe parameters from the model that contains all the covariates,
the gene, and the gene-environment interaction. The values has the same
requirements as the gonly value.}

\item{levene}{Logical array indicating which Levene tests to run. The
first value indicates to run the Levene test with the interaction
covariate and the second value indicates to run the Levene test without
the interaction covariate. This value may be a logical array of length
one or two. If only one value is passed it is used for both tests.}

\item{outfile}{The file name for the results Can be blank.
If the value is "", the results are returned as a data frame. Default
value is ""}

\item{skipfile}{The name of the file to write the SNPs that were not
used and the reason they weren't used. If the value is blank, there is
no output of the unused SNPs. Default value is "".}

\item{minmaf}{Minimum minor allele frequency of SNPs to include
in analysis. SNPS that have less than 20 minor alleles observed
will be excluded from the analysis regardless of the value of
minmaf. A value of 0 indicates to use all the SNPs that have 20
minor alleles observed. Default value is 0.}

\item{blksize}{Size of blocks of SNPs to read in at one time.
Larger blocks can improve overall speed but require larger
amounts of computer memory. A value of 0 indicates to use the
recommended block size. Default value is 0.}
}
\value{
0
}
\description{
Run a linear gweis using genetic data from a binary dosage or vcf file
}
\examples{
bdinfo <- readRDS(system.file("extdata/pdata_4_1.bdinfo", package = "GxEScanR"))
covdata <- readRDS(system.file("extdata/covdata.rds", package = "GxEScanR"))

results <- gweis(data = covdata, ginfo = bdinfo)
}
